+++ ovp_core/models/__init__.py
@@ -1,4 +1,4 @@
-from .address import GoogleAddress, AddressComponent, AddressComponentType, GoogleRegion
+from .address import GoogleAddress, AddressComponent, AddressComponentType, GoogleRegion, SimpleAddress
 from .cause import Cause
 from .skill import Skill
 from .lead import Lead