+++ ovp_core/models/address/__init__.py
@@ -1,0 +1,2 @@
+from .google_address import GoogleAddress, AddressComponent, AddressComponentType, GoogleRegion
+from .simple_address import SimpleAddress